handle the read only admin on the front end

+++ app/Http/Controllers/Institution/ManagementDataController.php
@@ -48,6 +48,7 @@ class ManagementDataController extends Controller
         $user = Auth::user();
         if ($user == null) return redirect('/login');
         $user->authorizeRoles('University Admin');
+        if ($user->read_only) return redirect("/institution/general");
 
         $institution = $user->institution();
         $managements = array();
@@ -89,6 +90,7 @@ class ManagementDataController extends Controller
         $user = Auth::user();
         if ($user == null) return redirect('/login');
         $user->authorizeRoles('University Admin');
+        if ($user->read_only) return redirect("/institution/general");
         $institution = $user->institution();
 
         $management_data = new ManagementData();
@@ -130,6 +132,7 @@ class ManagementDataController extends Controller
         $user = Auth::user();
         if ($user == null) return redirect('/login');
         $user->authorizeRoles('University Admin');
+        if ($user->read_only) return redirect("/institution/general");
 
         $institution = $user->institution();
         $managements = array();
@@ -173,7 +176,6 @@ class ManagementDataController extends Controller
         $user = Auth::user();
         if ($user == null) return redirect('/login');
         $user->authorizeRoles('University Admin');
-        $institution = $user->institution();
 
         $management_data = ManagementData::find($id);
         $management_data->required = $request->input('required_positions');
@@ -194,6 +196,11 @@ class ManagementDataController extends Controller
      */
     public function destroy($id)
     {
+        $user = Auth::user();
+        if ($user == null) return redirect('/login');
+        $user->authorizeRoles('University Admin');
+        if ($user->read_only) return redirect("/institution/general");
+
         $item = ManagementData::find($id);
         $item->delete();
         return redirect('institution/management-data')->with('primary', 'Successfully Deleted');
